Opengraph Metatags added for XrPages and special og:article metatags for BlogEntryPages. Images for pages are not scaled or modified but their height and width is stored in the opengraph tags. Used tags for pages and blogentries: og:site_name: constant "Extinction Rebellion Deutschland" (is there another source for this string?) og:locale: constant "de_DE" og:url: the absolute uri of the request og:image: only when set in wagtail, the url of that image og:image:width: if active, the width of the image og:image.height: if active, the height of the image og:description: if given in wagtail, the description text og:title: if given in wagtail in "veröffentlichen" this title is taken, else the normal title og:type: normally "website". "article" for BlogEntryPages
For BlogEntryPages only:
og:article:published_date: The date of the blog publishing
og:article:author:username: The Wagtail username of the author
og:article:section: The associated local group (either groupname or "Deutschland").

Improvment Ideas:
Maybe the titles of the Blogs should include "Blog" or "Blog OG" additionally to the title?
Or maybe the sections should be named "Blog" or "Blog of OG"?