Events are now date-filtered based on their dates.start entries
Instead of filtering events based on the EventPage's start_date and
end_date (which are processed on each save() of the EventPage by
iterating though it's 'dates'), we now use the associated dates.start
value alone to sort and filter the events. These EventDates are atomic
enough so we don't have to look at end dates which might not even exist.

EventDates that started yesterday are not shown in upcoming, although
they might end only today or in the future. I think thats fair enough.