Change the used dates for events to individual dates
Until now the start_date and end_date of an event has been derived from
the first and last entry in the 'dates' table.
This was useful for the case where you have an event spanning multiple
days because this event appeared as one event.
However, it has mainly be used for "repeating" events for which we don't
have (yet).
In this use case this leads to ugly and non-useful event date ordering
in the list and also each entry in the list doesn't show its real
underlying date.
With this commit the real singular dates of the event instances are used
for ordering and display of the event cards in the list.